requirements.txt
# Core Data Science
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.1.0
opencv-python==4.8.0.76
scikit-learn==1.3.2

# Deep Learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Jupyter and Visualization
jupyter==1.0.0
ipykernel==6.25.0
plotly==5.16.1
tqdm==4.66.1

# AWS and Data Access
boto3==1.28.25
awscli==1.29.25

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

gradio